Break large files apart, deal with edge cases.
