+++ include/albatross/PatchworkGP
@@ -17,6 +17,7 @@
 
 #include <albatross/src/utils/block_utils.hpp>
 #include <albatross/src/utils/eigen_utils.hpp>
+#include <albatross/src/models/patchwork_gp_details.hpp>
 #include <albatross/src/models/patchwork_gp.hpp>
 
 #endif
